+++ opennmt/optimizers/__init__.py
@@ -1,0 +1,7 @@
+"""Module defining custom optimizers."""
+
+from opennmt.optimizers.adafactor import AdafactorOptimizer
+from opennmt.optimizers.adafactor import get_optimizer_from_params \
+    as get_adafactor_optimizer_from_params
+
+from opennmt.optimizers.multistep_adam import MultistepAdamOptimizer