+++ opennmt/utils/evaluator.py
@@ -38,13 +38,15 @@ class ExternalEvaluator(object):
       self._summarize_score(step, score)
     self._log_score(score)
 
+  def _summarize_value(self, step, tag, value):
+    summary = tf.Summary(value=[tf.Summary.Value(tag=tag, simple_value=value)])
+    self._summary_writer.add_summary(summary, step)
+
   # Some evaluators may return several scores so let them the ability to
   # define how to log the score result.
 
   def _summarize_score(self, step, score):
-    summary = tf.Summary(value=[tf.Summary.Value(
-        tag="external_evaluation/{}".format(self.name()), simple_value=score)])
-    self._summary_writer.add_summary(summary, step)
+    self._summarize_value(step, "external_evaluation/{}".format(self.name()), score)
 
   def _log_score(self, score):
     tf.logging.info("%s evaluation score: %f", self.name(), score)
@@ -58,6 +60,35 @@ class ExternalEvaluator(object):
   def score(self, labels_file, predictions_path):
     """Scores the predictions against the true output labels."""
     raise NotImplementedError()
+
+
+class ROUGEEvaluator(ExternalEvaluator):
+  """ROUGE evaluator based on https://github.com/pltrdy/rouge."""
+
+  def __init__(self, labels_file=None, output_dir=None):
+    try:
+      import rouge  # pylint: disable=unused-variable
+    except ImportError:
+      raise ImportError("Please install the 'rouge' package, run 'pip install rouge==0.3.0'")
+    super(ROUGEEvaluator, self).__init__(labels_file=output_dir, output_dir=output_dir)
+
+  def name(self):
+    return "ROUGE"
+
+  def _summarize_score(self, step, score):
+    self._summarize_value(step, "external_evaluation/ROUGE-1", score["rouge-1"])
+    self._summarize_value(step, "external_evaluation/ROUGE-2", score["rouge-2"])
+    self._summarize_value(step, "external_evaluation/ROUGE-L", score["rouge-l"])
+
+  def _log_score(self, score):
+    tf.logging.info("Evaluation score: ROUGE-1 = %f; ROUGE-2 = %f; ROUGE-L = %s",
+                    score["rouge-1"], score["rouge-2"], score["rouge-l"])
+
+  def score(self, labels_file, predictions_path):
+    from rouge import FilesRouge
+    files_rouge = FilesRouge(predictions_path, labels_file)
+    rouge_scores = files_rouge.get_scores(avg=True)
+    return {k:v["f"] for k, v in six.iteritems(rouge_scores)}
 
 
 class BLEUEvaluator(ExternalEvaluator):
@@ -133,6 +164,8 @@ def external_evaluation_fn(evaluators_name, labels_file, output_dir=None):
       evaluator = BLEUEvaluator(labels_file=labels_file, output_dir=output_dir)
     elif name == "bleu-detok":
       evaluator = BLEUDetokEvaluator(labels_file=labels_file, output_dir=output_dir)
+    elif name == "rouge":
+      evaluator = ROUGEEvaluator(labels_file=labels_file, output_dir=output_dir)
     else:
       raise ValueError("No evaluator associated with the name: {}".format(name))
     evaluators.append(evaluator)